# Switch to xxhash3 for improved hash performance