--- a/packages/core/core/src/AssetGraph.js
+++ b/packages/core/core/src/AssetGraph.js
@@ -19,12 +19,8 @@ import type {
 } from './types';
 
 import invariant from 'assert';
-import crypto from 'crypto';
-import {
-  md5FromObject,
-  md5FromOrderedObject,
-  objectSortedEntries,
-} from '@parcel/utils';
+import {hashString, Hash} from '@parcel/hash';
+import {hashObject, objectSortedEntries} from '@parcel/utils';
 import nullthrows from 'nullthrows';
 import ContentGraph, {type SerializedContentGraph} from './ContentGraph';
 import {createDependency} from './Dependency';
@@ -57,15 +53,17 @@ export function nodeFromDep(dep: Dependency): DependencyNode {
 
 export function nodeFromAssetGroup(assetGroup: AssetGroup): AssetGroupNode {
   return {
-    id: md5FromOrderedObject({
-      filePath: assetGroup.filePath,
-      env: assetGroup.env.id,
-      isSource: assetGroup.isSource,
-      sideEffects: assetGroup.sideEffects,
-      code: assetGroup.code,
-      pipeline: assetGroup.pipeline,
-      query: assetGroup.query ? objectSortedEntries(assetGroup.query) : null,
-    }),
+    id: hashString(
+      assetGroup.filePath +
+        assetGroup.env.id +
+        String(assetGroup.isSource) +
+        String(assetGroup.sideEffects) +
+        (assetGroup.code ?? '') +
+        (assetGroup.pipeline ?? '') +
+        (assetGroup.query
+          ? JSON.stringify(objectSortedEntries(assetGroup.query))
+          : ''),
+    ),
     type: 'asset_group',
     value: assetGroup,
     usedSymbolsDownDirty: true,
@@ -93,7 +91,7 @@ export function nodeFromEntrySpecifier(entry: string): EntrySpecifierNode {
 
 export function nodeFromEntryFile(entry: Entry): EntryFileNode {
   return {
-    id: 'entry_file:' + md5FromObject(entry),
+    id: 'entry_file:' + hashObject(entry),
     type: 'entry_file',
     value: entry,
   };
@@ -569,18 +567,18 @@ export default class AssetGraph extends ContentGraph<AssetGraphNode> {
       return this.hash;
     }
 
-    let hash = crypto.createHash('md5');
+    let hash = new Hash();
     // TODO: sort??
     this.traverse(nodeId => {
       let node = nullthrows(this.getNode(nodeId));
       if (node.type === 'asset') {
-        hash.update(nullthrows(node.value.outputHash));
+        hash.writeString(nullthrows(node.value.outputHash));
       } else if (node.type === 'dependency' && node.value.target) {
-        hash.update(JSON.stringify(node.value.target));
+        hash.writeString(JSON.stringify(node.value.target));
       }
     });
 
-    this.hash = hash.digest('hex');
+    this.hash = hash.finish();
     return this.hash;
   }
 }