--- a/packages/core/core/src/Transformation.js
+++ b/packages/core/core/src/Transformation.js
@@ -22,11 +22,7 @@ import type {LoadedPlugin} from './ParcelConfig';
 
 import path from 'path';
 import nullthrows from 'nullthrows';
-import {
-  md5FromOrderedObject,
-  normalizeSeparators,
-  objectSortedEntries,
-} from '@parcel/utils';
+import {normalizeSeparators, objectSortedEntries} from '@parcel/utils';
 import logger, {PluginLogger} from '@parcel/logger';
 import {init as initSourcemaps} from '@parcel/source-map';
 import ThrowableDiagnostic, {
@@ -35,7 +31,7 @@ import ThrowableDiagnostic, {
   md,
 } from '@parcel/diagnostic';
 import {SOURCEMAP_EXTENSIONS} from '@parcel/utils';
-import crypto from 'crypto';
+import {hashString} from '@parcel/hash';
 
 import {createDependency} from './Dependency';
 import ParcelConfig from './ParcelConfig';
@@ -333,29 +329,26 @@ export default class Transformation {
   }
 
   async getPipelineHash(pipeline: Pipeline): Promise<string> {
-    let hash = crypto.createHash('md5');
+    let hashes = '';
     for (let transformer of pipeline.transformers) {
       let key = `${transformer.name}:${transformer.resolveFrom}`;
-      hash.update(
+      hashes +=
         this.request.devDeps.get(key) ??
-          this.devDepRequests.get(key)?.hash ??
-          '',
-      );
+        this.devDepRequests.get(key)?.hash ??
+        '';
 
       let config = this.configs.get(transformer.name);
       if (config) {
-        hash.update(
-          await getConfigHash(config, transformer.name, this.options),
-        );
+        hashes += await getConfigHash(config, transformer.name, this.options);
 
         for (let devDep of config.devDeps) {
           let key = `${devDep.moduleSpecifier}:${devDep.resolveFrom}`;
-          hash.update(nullthrows(this.devDepRequests.get(key)).hash);
+          hashes += nullthrows(this.devDepRequests.get(key)).hash;
         }
       }
     }
 
-    return hash.digest('hex');
+    return hashString(hashes);
   }
 
   async addDevDependency(
@@ -578,21 +571,23 @@ export default class Transformation {
     invalidationHash: string,
     pipelineHash: string,
   ): string {
-    let assetsKeyInfo = assets.map(a => ({
-      filePath: a.value.filePath,
-      pipeline: a.value.pipeline,
-      hash: a.value.hash,
-      uniqueKey: a.value.uniqueKey,
-      query: a.value.query ? objectSortedEntries(a.value.query) : '',
-    }));
+    let assetsKeyInfo = assets
+      .map(a => [
+        a.value.filePath,
+        a.value.pipeline,
+        a.value.hash,
+        a.value.uniqueKey,
+        a.value.query ? JSON.stringify(objectSortedEntries(a.value.query)) : '',
+      ])
+      .join('');
 
-    return md5FromOrderedObject({
-      parcelVersion: PARCEL_VERSION,
-      assets: assetsKeyInfo,
-      env: this.request.env,
-      invalidationHash,
-      pipelineHash,
-    });
+    return hashString(
+      PARCEL_VERSION +
+        assetsKeyInfo +
+        this.request.env.id +
+        invalidationHash +
+        pipelineHash,
+    );
   }
 
   async loadPipeline(