--- a/packages/core/core/src/UncommittedAsset.js
+++ b/packages/core/core/src/UncommittedAsset.js
@@ -25,13 +25,13 @@ import {Readable} from 'stream';
 import SourceMap from '@parcel/source-map';
 import {
   bufferStream,
-  md5FromString,
   blobToStream,
   streamFromPromise,
   TapStream,
   loadSourceMap,
   SOURCEMAP_RE,
 } from '@parcel/utils';
+import {hashString} from '@parcel/hash';
 import {createDependency, mergeDependencies} from './Dependency';
 import {mergeEnvironments} from './Environment';
 import {PARCEL_VERSION} from './constants';
@@ -129,12 +129,10 @@ export default class UncommittedAsset {
     this.value.contentKey = contentKey;
     this.value.mapKey = mapKey;
     this.value.astKey = astKey;
-    this.value.outputHash = md5FromString(
-      [
-        this.value.hash,
-        pipelineKey,
-        await getInvalidationHash(this.getInvalidations(), this.options),
-      ].join(':'),
+    this.value.outputHash = hashString(
+      (this.value.hash ?? '') +
+        pipelineKey +
+        (await getInvalidationHash(this.getInvalidations(), this.options)),
     );
 
     if (this.content != null) {
@@ -304,7 +302,7 @@ export default class UncommittedAsset {
   }
 
   getCacheKey(key: string): string {
-    return md5FromString(
+    return hashString(
       PARCEL_VERSION + key + this.value.id + (this.value.hash || ''),
     );
   }