--- a/packages/core/core/src/assetUtils.js
+++ b/packages/core/core/src/assetUtils.js
@@ -77,7 +77,9 @@ export function createAssetIdFromOptions(options: AssetOptions): string {
       options.type +
       options.env.id +
       uniqueKey +
+      ':' +
       (options.pipeline ?? '') +
+      ':' +
       queryString,
   );
 }