--- a/packages/core/core/src/requests/AssetRequest.js
+++ b/packages/core/core/src/requests/AssetRequest.js
@@ -55,7 +55,9 @@ function getId(input: AssetRequestInput) {
       String(input.isSource) +
       String(input.sideEffects) +
       (input.code ?? '') +
+      ':' +
       (input.pipeline ?? '') +
+      ':' +
       (input.query ? JSON.stringify(objectSortedEntries(input.query)) : ''),
   );
 }