--- a/packages/utils/hash/build.rs
+++ b/packages/utils/hash/build.rs
@@ -1,0 +1,5 @@
+extern crate napi_build;
+
+fn main() {
+  napi_build::setup();
+}