T-961

Parcel relies heavily on hashing for many things, including cache keys, content hashes, etc. We've used md5 for this so far since it is a built in hash function in Node.js. However, it shows up pretty high in profiles of overall performance. This PR replaces md5 with [xxhash](http://cyan4973.github.io/xxHash/), which is one of the fastest hash functions out there.

Specifically, this is an napi wrapper around [this Rust implementation](https://github.com/DoumanAsh/xxhash-rust) of xxhash. There were some existing npm modules available for this, but most were either unmaintained or only accepted Buffers and not strings. This offered some performance increase, but the overhead of converting between strings and buffers was too high.

This benchmark shows the difference between md5, xxhash-addon (an existing npm module that only accepts buffers), and xxhash-rust.

```
md5 209,999.15 opts/sec (mean: 0.005ms, stddev: 0.029ms, 500 samples)
xxhash-addon 415,511.89 opts/sec (mean: 0.002ms, stddev: 0.003ms, 500 samples)
xxhash-rust 1,517,985.09 opts/sec (mean: 658.768ns, stddev: 996.69ns, 500 samples)
```

Overall this improves performance on the esbuild benchmark by ~5%.