@devongovett 2021-05-29
One thing to consider: xxhash3 is a 64-bit hash, whereas md5 is a 128-bit hash. There is also a 128-bit version of xxhash that we could consider, though it is slightly slower. I'm not sure if this matters in practice. A 64-bit hash should probably give us a low enough probability of collision for our uses I think.