@mischnic 2021-05-30
Can you leave the benchmark script in there? [`d003549` (#6373)](https://github.com/parcel-bundler/parcel/pull/6373/commits/d0035492bf8fb9da61d08a4f6b721fba852d3321#diff-89d72a5991f686f52f33d685f11aa52bea711090ec139b133adfc08cf43e3841)

Might be interesting for the future or the wasm version